Implement generic rerouting
The rerouting now works for an arbitrary number of trains at
arbitrary stations via an arbitrary station. The information of
the trains to be rerouted needs to be specified in the file
`configure/data/trains-to-reroute.csv`. Each row in this file
should contain four columns:
"trainid,reroute_start_station,reroute_via_station,reroute_end_station"